Add styled content steps
